changes from me - changed: mysql is now working with ado.net assembly. also changed dolconfig to don't use tinyint(1) as boolean since it causes problems. - changed: npctemplage ghost is now renamed flags to allow various flags to be applyed to npctemplates - fixed: vampiir speed cast while in combat 'll not longer draw mana to fail - changed: voucher mount model are now using npctemplate, but compatibility is preserved. You can use these npctemplate, do not forget to ensure Ghost column is renamed Flags and changed from tinyint 1 to tinyint 3 insert into `npctemplate` (`NpcTemplate_ID`, `TemplateId`, `Name`, `GuildName`, `Model`, `Size`, `MaxSpeed`, `EquipmentTemplateID`, `Flags`, `MeleeDamageType`, `ParryChance`, `EvadeChance`, `BlockChance`, `LeftHandSwingChance`, `Spells`, `Styles`, `Strength`, `Constitution`, `Dexterity`, `Quickness`, `Intelligence`, `Piety`, `Charisma`, `Empathy`, `Abilities`, `AggroLevel`, `AggroRange`, `ClassType`, `Level`, `BodyType`, `MaxDistance`, `TetherRange`, `PackageID`) values('Voucher mount - horse','300','Horse',NULL,'448;449;450;2172;2173;2174','63','600',NULL,'0','0','0','0','0','0',NULL,NULL,'0','0','0','0','0','0','0','0',NULL,'0','0','','55',NULL,NULL,NULL,'Voucher mounts'); insert into `npctemplate` (`NpcTemplate_ID`, `TemplateId`, `Name`, `GuildName`, `Model`, `Size`, `MaxSpeed`, `EquipmentTemplateID`, `Flags`, `MeleeDamageType`, `ParryChance`, `EvadeChance`, `BlockChance`, `LeftHandSwingChance`, `Spells`, `Styles`, `Strength`, `Constitution`, `Dexterity`, `Quickness`, `Intelligence`, `Piety`, `Charisma`, `Empathy`, `Abilities`, `AggroLevel`, `AggroRange`, `ClassType`, `Level`, `BodyType`, `MaxDistance`, `TetherRange`, `PackageID`) values('Voucher mount - dragonfly','301','Dragon Fly',NULL,'1207','63','1120',NULL,'32','0','0','0','0','0',NULL,NULL,'0','0','0','0','0','0','0','0',NULL,'0','0','','55',NULL,NULL,NULL,'Voucher mounts'); insert into `npctemplate` (`NpcTemplate_ID`, `TemplateId`, `Name`, `GuildName`, `Model`, `Size`, `MaxSpeed`, `EquipmentTemplateID`, `Flags`, `MeleeDamageType`, `ParryChance`, `EvadeChance`, `BlockChance`, `LeftHandSwingChance`, `Spells`, `Styles`, `Strength`, `Constitution`, `Dexterity`, `Quickness`, `Intelligence`, `Piety`, `Charisma`, `Empathy`, `Abilities`, `AggroLevel`, `AggroRange`, `ClassType`, `Level`, `BodyType`, `MaxDistance`, `TetherRange`, `PackageID`) values('Voucher mount - wyvern','302','Ampheretere',NULL,'1235','63','1120',NULL,'32','0',NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,'','55',NULL,NULL,NULL,'Voucher mounts'); insert into `npctemplate` (`NpcTemplate_ID`, `TemplateId`, `Name`, `GuildName`, `Model`, `Size`, `MaxSpeed`, `EquipmentTemplateID`, `Flags`, `MeleeDamageType`, `ParryChance`, `EvadeChance`, `BlockChance`, `LeftHandSwingChance`, `Spells`, `Styles`, `Strength`, `Constitution`, `Dexterity`, `Quickness`, `Intelligence`, `Piety`, `Charisma`, `Empathy`, `Abilities`, `AggroLevel`, `AggroRange`, `ClassType`, `Level`, `BodyType`, `MaxDistance`, `TetherRange`, `PackageID`) values('Voucher mount - griffin','303','Gryphon',NULL,'1236','63','1120',NULL,'32','0',NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,NULL,'','55',NULL,NULL,NULL,'Voucher mounts');
changes from Dunnerholl
- gamestablemaster is no longer parsing ticket item to retrieve mount model

changes from Xali
- Spells can no longer be interrupted once they are over 2/3 cast time
- Dwarf berserker got correct model

+++ GameServer/gameutils/INpcTemplate.cs
@@ -63,7 +63,7 @@ namespace DOL.GS
 		/// <summary>
 		/// Gets the template npc flags
 		/// </summary>
-		uint Flags { get; }
+		byte Flags { get; }
 
 		/// <summary>
 		/// Gets the template npc inventory
